Move stepper to separate folder
