+++ client/tasks/fills/tax/utils.ts
@@ -1,4 +1,9 @@
 /**
+ * External dependencies
+ */
+import { getAdminLink } from '@woocommerce/wc-admin-settings';
+
+/**
  * Plugins required to automate taxes.
  */
 export const AUTOMATION_PLUGINS = [ 'jetpack', 'woocommerce-services' ];
@@ -15,4 +20,13 @@ export const hasCompleteAddress = ( generalSettings ): boolean => {
 		woocommerce_store_postcode: storePostCode,
 	} = generalSettings;
 	return Boolean( storeAddress && defaultCountry && storePostCode );
+};
+
+/**
+ * Redirect to the core tax settings screen.
+ */
+export const redirectToTaxSettings = (): void => {
+	window.location = getAdminLink(
+		'admin.php?page=wc-settings&tab=tax&section=standard&wc_onboarding_active_task=tax'
+	);
 };
